+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `getFeatures()` was renamed to be less ambiguous and better align with
   internal SomaLogic code usage. Now use `getAnalytes()`
+* `getFeatureData()` was also renamed to `getAnalyteInfo()`
 
 # SomaDataIO 5.0.0